requirements.txt
%%writefile requirements.txt
tensorflow
streamlit
pillow
numpy
pandas
gdown
